리팩토링 수행 - 1
