+++ runner/driver/tr-driver/tr-driver.c
@@ -16,9 +16,6 @@
 #include <sys/types.h>
 #include <sys/wait.h>
 #include <sys/syscall.h>
-#include <sys/ipc.h>
-#include <sys/shm.h>
-#include <sys/sem.h>
 
 /**< external header */
 #include <json.h>
@@ -68,40 +65,6 @@ static void tr_kill_handle(int signum)
         }
         _exit(EXIT_SUCCESS);
 }
-
-#ifdef TR_DEBUG
-/**
- * @brief 현재 info의 내용을 출력합니다. 
- *
- * @param info 출력하고자 하는 info의 포인터에 해당합니다.
- */
-static void tr_debug(const struct tr_info *info)
-{
-        pr_info(INFO,
-                "\n"
-                "\t\t[[ current %p ]]\n"
-                "\t\tpid: %d\n"
-                "\t\ttime: %u\n"
-                "\t\tq_depth: %u\n"
-                "\t\tnr_thread: %u\n"
-                "\t\tweight: %u\n"
-                "\t\tqid: %d\n"
-                "\t\tshmid: %d\n"
-                "\t\tsemid: %d\n"
-                "\t\tprefix_cgroup_name: %s\n"
-                "\t\tscheduler: %s\n"
-                "\t\tname: %s\n"
-                "\t\ttrace_data_path: %s\n"
-                "\t\tdevice: %s\n"
-                "\t\tglobal_config: %p\n"
-                "\t\tnext: %p\n",
-                info, info->pid, info->time, info->q_depth, info->nr_thread,
-                info->weight, info->qid, info->shmid, info->semid,
-                info->prefix_cgroup_name, info->scheduler, info->cgroup_id,
-                info->trace_data_path, info->device, info->global_config,
-                info->next);
-}
-#endif
 
 /**
  * @brief 실질적으로 trace-replay 관련 구조체의 동적 할당된 내용을 해제하는 부분에 해당합니다.
@@ -155,62 +118,13 @@ static void __tr_free(void)
 }
 
 /**
- * @brief 정수 형태의 `info->(member)`에 json에서 값을 읽어서 값을 주도록 합니다. 
- *
- * @param setting 탐색을 할 특정 위치를 지칭합니다.
- * @param key json에서 가져오고자 하는 데이터의 key 또는 field에 해당합니다.
- * @param member 실제 값이 삽입되는 위치에 해당합니다.
- * @param is_print 에러를 사용자에게 출력을 할 지 말지를 선택합니다.
- *
- * @return 성공적으로 입력이 되었다면 0이 반환되고, 그렇지 않은 경우에는 -EINVAL이 반환됩니다.
- */
-static int tr_info_int_value_set(struct json_object *setting, const char *key,
-                                 unsigned int *member, int is_print)
-{
-        struct json_object *tmp = NULL;
-        if (!json_object_object_get_ex(setting, key, &tmp)) {
-                if (TR_ERROR_PRINT == is_print) {
-                        pr_info(ERROR, "Not exist error (key: %s)\n", key);
-                }
-                return -EINVAL;
-        }
-        *member = json_object_get_int(tmp);
-        return 0;
-}
-
-/**
- * @brief 문자열 형태의 `info->(member)`에 json에서 값을 읽어서 값을 주도록 합니다. 
- *
- * @param setting 탐색을 할 특정 위치를 지칭합니다.
- * @param key json에서 가져오고자 하는 데이터의 key 또는 field에 해당합니다.
- * @param member 실제 값이 삽입되는 위치에 해당합니다.
- * @param is_print 에러를 사용자에게 출력을 할 지 말지를 선택합니다.
- *
- * @return 성공적으로 입력이 되었다면 0이 반환되고, 그렇지 않은 경우에는 -EINVAL이 반환됩니다.
- */
-static int tr_info_str_value_set(struct json_object *setting, const char *key,
-                                 char *member, int is_print)
-{
-        struct json_object *tmp = NULL;
-        if (!json_object_object_get_ex(setting, key, &tmp)) {
-                if (TR_ERROR_PRINT == is_print) {
-                        pr_info(ERROR, "Not exist error (key: %s)\n", key);
-                }
-                return -EINVAL;
-        }
-        strcpy(member, json_object_get_string(tmp));
-        generic_strip_string(member, '\"');
-        return 0;
-}
-
-/**
  * @brief 현재 입력되는 scheduler가 driver가 지원하는 지를 확인하도록 합니다.
  *
  * @param scheduler 검사하고자 하는 스케쥴러 문자열을 가진 문자열 포인터입니다.
  *
  * @return 가지고 있는 경우 0이 반환되고, 그렇지 않은 경우 -EINVAL이 반환됩니다.
  */
-static int valid_scheduler_test(const char *scheduler)
+int tr_valid_scheduler_test(const char *scheduler)
 {
         const char *index = tr_valid_scheduler[0];
         while (index != NULL) {
@@ -220,165 +134,6 @@ static int valid_scheduler_test(const char *scheduler)
                 index++;
         }
         return -EINVAL;
-}
-
-/**
- * @brief 각각의 프로세스의 동작 옵션을 설정을 해주도록 합니다.
- *
- * @param setting 설정 값을 가지고 있는 json 객체의 포인터에 해당합니다.
- * @param index json의 task_option의 배열 순번에 대한 정보를 가집니다.
- * @param info 실제 값이 들어가는 위치에 해당합니다.
- *
- * @return 정상적으로 초기화가 된 경우 0을 그렇지 않은 경우 적절한 오류 번호를 반환합니다. 
- */
-static int __tr_info_init(struct json_object *setting, int index,
-                          struct tr_info *info)
-{
-        struct json_object *tmp;
-        int ret = 0;
-
-        ENTRY item; /**< hsearch를 위해서 사용되는 변수입니다. */
-        ENTRY *result;
-
-        assert(NULL != info);
-
-        if (!json_object_object_get_ex(setting, "task_option", &tmp)) {
-                pr_info(ERROR, "Not exist error (key: %s)\n", "task_option");
-                ret = -EINVAL;
-                goto exception;
-        }
-
-        tmp = json_object_array_get_idx(tmp, index);
-        if (NULL == tmp) {
-                pr_info(ERROR, "Array out-of-bound error (index: %d)\n", index);
-                ret = -EINVAL;
-                goto exception;
-        }
-
-        tr_info_int_value_set(tmp, "time", &info->time, TR_PRINT_NONE);
-        tr_info_int_value_set(tmp, "q_depth", &info->q_depth, TR_PRINT_NONE);
-        tr_info_int_value_set(tmp, "nr_thread", &info->nr_thread,
-                              TR_PRINT_NONE);
-        tr_info_str_value_set(tmp, "prefix_cgroup_name",
-                              info->prefix_cgroup_name, TR_PRINT_NONE);
-        tr_info_str_value_set(tmp, "scheduler", info->scheduler, TR_PRINT_NONE);
-        tr_info_str_value_set(tmp, "trace_replay_path", info->trace_replay_path,
-                              TR_PRINT_NONE);
-        tr_info_str_value_set(tmp, "device", info->device, TR_PRINT_NONE);
-        ret = valid_scheduler_test(info->scheduler);
-        if (0 != ret) {
-                pr_info(ERROR, "Unsupported scheduler (name: %s)\n",
-                        info->scheduler);
-                goto exception;
-        }
-
-        ret = tr_info_int_value_set(tmp, "weight", &info->weight,
-                                    TR_ERROR_PRINT);
-        if (0 != ret) {
-                goto exception;
-        }
-
-        ret = tr_info_str_value_set(tmp, "trace_data_path",
-                                    info->trace_data_path, TR_ERROR_PRINT);
-        if (0 != ret) {
-                goto exception;
-        }
-
-        ret = tr_info_str_value_set(tmp, "cgroup_id", info->cgroup_id,
-                                    TR_ERROR_PRINT);
-        if (0 != ret) {
-                goto exception;
-        }
-
-        if (-1 == (ret = access(info->trace_data_path, F_OK))) {
-                pr_info(ERROR, "Trace data file not exist: %s\n",
-                        info->trace_data_path);
-                goto exception;
-        } else {
-                pr_info(INFO, "Trace data file exist: %s\n",
-                        info->trace_data_path);
-        }
-        info->ppid = getpid();
-
-        item.key = info->cgroup_id;
-        if (NULL != (result = hsearch(item, FIND))) {
-                pr_info(ERROR, "Duplicate c-group name detected (name: %s)\n",
-                        result->key);
-                ret = -EINVAL;
-                goto exception;
-        }
-
-        item.key = info->cgroup_id;
-        item.data = info;
-        hsearch(item, ENTER);
-
-exception:
-        return ret;
-}
-
-/**
- * @brief 각 프로세스에 들어가는 info 객체를 __생성__하고 구축하여 반환합니다.
- *
- * @param setting 설정 값을 가지고 있는 json 객체의 포인터에 해당합니다.
- * @param index json의 task_option의 배열 순번에 대한 정보를 가집니다.
- *
- * @return 정상적으로 초기화가 된 경우 0을 그렇지 않은 경우 적절한 오류 번호를 반환합니다. 
- */
-static struct tr_info *tr_info_init(struct json_object *setting, int index)
-{
-        struct tr_info *info;
-        int ret = 0;
-
-        info = (struct tr_info *)malloc(sizeof(struct tr_info));
-        if (!info) {
-                pr_info(ERROR, "Memory allocation fail (name: %s)\n", "info");
-                ret = -ENOMEM;
-                goto exception;
-        }
-
-        memset(info, 0, sizeof(struct tr_info));
-
-        ret |= tr_info_int_value_set(setting, "time", &info->time,
-                                     TR_ERROR_PRINT);
-        ret |= tr_info_int_value_set(setting, "q_depth", &info->q_depth,
-                                     TR_ERROR_PRINT);
-        ret |= tr_info_int_value_set(setting, "nr_thread", &info->nr_thread,
-                                     TR_ERROR_PRINT);
-        ret |= tr_info_str_value_set(setting, "prefix_cgroup_name",
-                                     info->prefix_cgroup_name, TR_ERROR_PRINT);
-        ret |= tr_info_str_value_set(setting, "scheduler", info->scheduler,
-                                     TR_ERROR_PRINT);
-        ret |= tr_info_str_value_set(setting, "device", info->device,
-                                     TR_ERROR_PRINT);
-        ret |= tr_info_str_value_set(setting, "trace_replay_path",
-                                     info->trace_replay_path, TR_ERROR_PRINT);
-        if (0 != ret) {
-                pr_info(ERROR, "error detected (errno: %d)\n", ret);
-                goto exception;
-        }
-
-        ret = valid_scheduler_test(info->scheduler);
-        if (0 != ret) {
-                pr_info(ERROR, "Unsupported scheduler (name: %s)\n",
-                        info->scheduler);
-                goto exception;
-        }
-
-        ret = __tr_info_init(setting, index, info);
-        if (0 != ret) {
-                pr_info(ERROR, "error detected (errno: %d)\n", ret);
-                goto exception;
-        }
-
-        info->next = NULL;
-
-        return info;
-exception:
-        if (NULL != info) {
-                free(info);
-                info = NULL;
-        }
-        return info;
 }
 
 /**
@@ -610,321 +365,6 @@ exit:
         return ret;
 }
 
-static struct json_object *tr_info_serializer(const struct tr_info *info)
-{
-        struct json_object *meta;
-
-        assert(NULL != info);
-        meta = json_object_new_object();
-        json_object_object_add(meta, "pid", json_object_new_int(info->pid));
-        json_object_object_add(meta, "time", json_object_new_int(info->time));
-        json_object_object_add(meta, "q_depth",
-                               json_object_new_int(info->q_depth));
-        json_object_object_add(meta, "nr_thread",
-                               json_object_new_int(info->nr_thread));
-        json_object_object_add(meta, "weight",
-                               json_object_new_int(info->weight));
-        json_object_object_add(meta, "qid", json_object_new_int(info->qid));
-        json_object_object_add(meta, "shmid", json_object_new_int(info->shmid));
-        json_object_object_add(meta, "semid", json_object_new_int(info->semid));
-        json_object_object_add(
-                meta, "prefix_cgroup_name",
-                json_object_new_string(info->prefix_cgroup_name));
-        json_object_object_add(meta, "scheduler",
-                               json_object_new_string(info->scheduler));
-        json_object_object_add(meta, "cgroup_id",
-                               json_object_new_string(info->cgroup_id));
-        json_object_object_add(meta, "trace_data_path",
-                               json_object_new_string(info->trace_data_path));
-        json_object_object_add(meta, "device", json_object_new_int(info->pid));
-
-        return meta;
-}
-
-static struct json_object *
-tr_realtime_log_serializer(const struct realtime_log *log)
-{
-        struct json_object *data;
-
-        data = json_object_new_object();
-        json_object_object_add(data, "type", json_object_new_int(log->type));
-        json_object_object_add(data, "time", json_object_new_double(log->time));
-        json_object_object_add(data, "remaining",
-                               json_object_new_double(log->remaining));
-        json_object_object_add(
-                data, "remaining_percentage",
-                json_object_new_double(log->remaining_percentage));
-        json_object_object_add(data, "avg_bw",
-                               json_object_new_double(log->avg_bw));
-        json_object_object_add(data, "cur_bw",
-                               json_object_new_double(log->cur_bw));
-        json_object_object_add(data, "lat", json_object_new_double(log->lat));
-        json_object_object_add(data, "time_diff",
-                               json_object_new_double(log->time_diff));
-
-        assert(NULL != log);
-        return data;
-}
-
-/**
- * @brief json으로 구조체에 있는 내용을 변환합니다.
- *
- * @param info 현재 tr_info 구조체의 포인터입니다.
- * @param log 현재 출력 로그에 해당합니다.
- * @param buffer 수행 결과 만들어진 json 내용이 들어가는 부분에 해당합니다.
- *
- * @warning buffer는 반드시 이 함수를 부르기 전에 할당되어야하며, 그 크기는 INTERVAL_RESULT_STRING_SIZE 보다 커야 합니다.
- * @todo 리팩토링이 필요합니다.
- */
-static void tr_realtime_serializer(const struct tr_info *info,
-                                   const struct realtime_log *log, char *buffer)
-{
-        struct json_object *object;
-
-        assert(NULL != info);
-        assert(NULL != log);
-        assert(NULL != buffer);
-
-        object = json_object_new_object();
-        json_object_object_add(object, "meta", tr_info_serializer(info));
-        json_object_object_add(object, "data", tr_realtime_log_serializer(log));
-
-        strcpy(buffer, json_object_to_json_string(object));
-
-        json_object_put(object);
-}
-
-static struct json_object *tr_total_trace_serializer(const struct trace *traces)
-{
-        struct json_object *_trace;
-        _trace = json_object_new_object();
-        json_object_object_add(_trace, "start_partition",
-                               json_object_new_double(traces->start_partition));
-        json_object_object_add(_trace, "total_size",
-                               json_object_new_double(traces->total_size));
-        json_object_object_add(_trace, "start_page",
-                               json_object_new_int64(traces->start_page));
-        json_object_object_add(_trace, "total_pages",
-                               json_object_new_int64(traces->total_pages));
-        return _trace;
-}
-
-static struct json_object *
-tr_total_config_serializer(const struct total_results *total,
-                           struct tr_total_json_object *jobject)
-{
-        struct json_object *config;
-        struct json_object *traces;
-        struct json_object **trace = jobject->trace;
-        int i;
-
-        assert(NULL != total);
-        assert(NULL != trace);
-
-        config = json_object_new_object();
-        json_object_object_add(config, "qdepth",
-                               json_object_new_int(total->config.qdepth));
-        json_object_object_add(config, "timeout",
-                               json_object_new_double(total->config.timeout));
-        json_object_object_add(config, "nr_trace",
-                               json_object_new_int(total->config.nr_trace));
-        json_object_object_add(config, "nr_thread",
-                               json_object_new_int(total->config.nr_thread));
-        json_object_object_add(config, "per_thread",
-                               json_object_new_int(total->config.per_thread));
-        json_object_object_add(
-                config, "result_file",
-                json_object_new_string(total->config.result_file));
-
-        traces = json_object_new_array();
-
-        for (i = 0; i < total->config.nr_thread; i++) {
-                trace[i] = tr_total_trace_serializer(&total->config.traces[i]);
-                json_object_array_add(traces, trace[i]);
-        }
-        json_object_object_add(config, "traces", traces);
-        return config;
-}
-
-static struct json_object *
-tr_synthetic_serializer(const struct synthetic *_synthetic)
-{
-        struct json_object *synthetic;
-        synthetic = json_object_new_object();
-        json_object_object_add(
-                synthetic, "working_set_size",
-                json_object_new_int(_synthetic->working_set_size));
-        json_object_object_add(synthetic, "utilization",
-                               json_object_new_int(_synthetic->utilization));
-        json_object_object_add(
-                synthetic, "touched_working_set_size",
-                json_object_new_int(_synthetic->touched_working_set_size));
-        json_object_object_add(synthetic, "io_size",
-                               json_object_new_int(_synthetic->io_size));
-        return synthetic;
-}
-
-static struct json_object *tr_stats_serializer(const struct trace_stat *_stats)
-{
-        struct json_object *stats;
-        struct tr_json_field *field, *begin, *end;
-        struct tr_json_field fields[] = {
-                { "exec_time", &_stats->exec_time },
-                { "avg_lat", &_stats->avg_lat },
-                { "avg_lat_var", &_stats->avg_lat_var },
-                { "lat_min", &_stats->lat_min },
-                { "lat_max", &_stats->lat_max },
-                { "iops", &_stats->iops },
-                { "total_bw", &_stats->total_bw },
-                { "read_bw", &_stats->read_bw },
-                { "write_bw", &_stats->write_bw },
-                { "total_traffic", &_stats->total_traffic },
-                { "read_traffic", &_stats->read_traffic },
-                { "write_traffic", &_stats->write_traffic },
-                { "read_ratio", &_stats->read_ratio },
-                { "total_avg_req_size", &_stats->total_avg_req_size },
-                { "read_avg_req_size", &_stats->read_avg_req_size },
-                { "write_avg_req_size", &_stats->write_avg_req_size },
-        };
-
-        stats = json_object_new_object();
-        begin = &fields[0];
-        end = &fields[sizeof(fields) / sizeof(struct tr_json_field)];
-        tr_json_field_traverse(field, begin, end)
-        {
-                struct json_object *current;
-                double value;
-
-                assert(NULL != field);
-                value = *(double *)field->member;
-
-                current = json_object_new_double(value);
-                assert(NULL != current);
-
-                json_object_object_add(stats, field->name, current);
-        }
-        return stats;
-}
-
-static struct json_object *
-tr_total_per_trace_serializer(const struct total_results *total,
-                              struct tr_total_json_object *jobject)
-{
-        struct json_object *per_traces;
-        struct json_object **synthetic = jobject->synthetic;
-        struct json_object **stats = jobject->stats;
-        struct json_object **per_trace = jobject->per_trace;
-        int i;
-
-        assert(NULL != total);
-        assert(NULL != synthetic);
-        assert(NULL != stats);
-        assert(NULL != per_trace);
-
-        per_traces = json_object_new_array();
-        for (i = 0; i < total->config.nr_thread; i++) {
-                struct json_object *_per_trace;
-                _per_trace = json_object_new_object();
-                json_object_object_add(
-                        _per_trace, "name",
-                        json_object_new_string(
-                                total->results.per_trace[i].name));
-                json_object_object_add(
-                        _per_trace, "issynthetic",
-                        json_object_new_int(
-                                total->results.per_trace[i].issynthetic));
-
-                synthetic[i] = tr_synthetic_serializer(
-                        &total->results.per_trace[i].synthetic);
-                json_object_object_add(_per_trace, "synthetic", synthetic[i]);
-
-                stats[i] =
-                        tr_stats_serializer(&total->results.per_trace[i].stats);
-                json_object_object_add(_per_trace, "stats", stats[i]);
-
-                json_object_object_add(
-                        _per_trace, "trace_reset_count",
-                        json_object_new_int(
-                                total->results.per_trace[i].trace_reset_count));
-
-                per_trace[i] = _per_trace;
-                json_object_array_add(per_traces, per_trace[i]);
-        }
-        return per_traces;
-}
-
-static struct json_object *
-tr_total_aggr_serializer(const struct total_results *total)
-{
-        return tr_stats_serializer(&total->results.aggr_result.stats);
-}
-
-static struct json_object *
-tr_total_result_serializer(const struct total_results *total,
-                           struct tr_total_json_object *jobject)
-{
-        struct json_object *results;
-        results = json_object_new_object();
-        json_object_object_add(results, "per_trace",
-                               tr_total_per_trace_serializer(total, jobject));
-
-        json_object_object_add(results, "aggr_result",
-                               tr_total_aggr_serializer(total));
-        return results;
-}
-
-static struct json_object *
-tr_total_results_serializer(const struct total_results *total,
-                            struct tr_total_json_object *jobject)
-{
-        struct json_object *total_results;
-        total_results = json_object_new_object();
-        json_object_object_add(total_results, "config",
-                               tr_total_config_serializer(total, jobject));
-        json_object_object_add(total_results, "results",
-                               tr_total_result_serializer(total, jobject));
-        return total_results;
-}
-
-/**
- * @brief json으로 구조체에 있는 내용을 변환합니다.
- *
- * @param info 현재 tr_info 구조체의 포인터입니다.
- * @param results 현재 info에 해당하는 전체 내용에 해당합니다.
- * @param buffer 수행 결과 만들어진 json 내용이 들어가는 부분에 해당합니다.
- *
- * @warning buffer는 반드시 이 함수를 부르기 전에 할당되어야하며, 그 크기는 TOTAL_RESULT_STRING_SIZE 보다 커야 합니다.
- * @todo 리팩토링이 필요합니다.
- */
-static void tr_total_serializer(const struct tr_info *info,
-                                const struct total_results *total, char *buffer)
-{
-        struct json_object *object;
-        struct json_object *info_object, *total_object;
-        struct tr_total_json_object *jobject;
-
-        assert(NULL != info);
-        assert(NULL != total);
-
-        jobject = (struct tr_total_json_object *)malloc(
-                sizeof(struct tr_total_json_object));
-        assert(NULL != jobject);
-
-        object = json_object_new_object();
-
-        info_object = tr_info_serializer(info);
-        total_object = tr_total_results_serializer(total, jobject);
-
-        json_object_object_add(object, "meta", info_object);
-        pr_info(INFO, "Adds meta success %p\n", (void *)info_object);
-
-        json_object_object_add(object, "data", total_object);
-        pr_info(INFO, "Adds data success %p\n", (void *)total_object);
-
-        strcpy(buffer, json_object_to_json_string(object));
-        json_object_put(object);
-        free(jobject);
-}
 /**
  * @brief trace-replay가 실행 중일 때의 정보를 반환합니다.
  *
